all case classes verified

+++ src/main/scala/models/Countries.scala
@@ -11,12 +11,12 @@ object Countries{
   def apply(line : String): Countries = {
     val data = line.split("," )
     new Countries(
-      data(0)
-      data(1)
-      data(2)
-      data(3)
-      data(4)
-      data(5)
+      data(0).toInt,
+      data(1),
+      data(2),
+      data(3),
+      data(4),
+      data(5),
     )
   }
 }
